DP-7022 fixed linting errors
