+++ package.json
@@ -26,13 +26,13 @@
     "preinstall": "yarn clean:generated",
     "postinstall": "lerna link",
     "lint": "yarn lint:css & yarn lint:js",
-    "lint:css": "stylelint './packages/**/*Style.ts'",
-    "lint:js": "eslint ./packages/asc-ui/src ./packages/asc-core/src --ext .ts,.tsx",
+    "lint:css": "stylelint \"./packages/**/*Style.ts\"",
+    "lint:js": "eslint \"./packages/asc-ui/src\" \"./packages/asc-core/src\" --ext .ts,.tsx",
     "prepare": "yarn build",
     "publish:packages": "yarn lint && yarn test && lerna publish",
     "start": "yarn build:watch & yarn storybook",
     "storybook": "start-storybook -p 6006 -c .storybook -s ./packages/asc-assets/lib/assets",
-    "stylelint": "stylelint ./packages/**/*Style.ts",
+    "stylelint": "stylelint \"./packages/**/*Style.ts\"",
     "test": "yarn jest",
     "test:coverage": "yarn jest --collect-coverage",
     "test:watch": "cross-env NODE_ENV=test jest --watch --onlyChanged --collect-coverage --verbose true",
