+++ packages/asc-ui/src/components/ContextMenu/types.ts
@@ -1,19 +1,8 @@
-export enum Position {
+enum Position {
   bottom = 'bottom',
   top = 'top',
 }
 
-export type ContextMenuWrapperStyleProps = {}
-
-export type ContextMenuItemProps = {
-  role?: string
-  id: string
-  label: string
-  focused: boolean
-  divider?: boolean
-}
-
-export type ContextMenuListProps = {
-  labelId?: string
-  position?: Position
+export default {
+  Position,
 }
